# Use Maybe a for vector and matrix products to improve runtime and space complexity