--- a/Math/LinearAlgebra/Sparse/IntMapUtilities.hs
+++ b/Math/LinearAlgebra/Sparse/IntMapUtilities.hs
@@ -10,8 +10,11 @@ import Data.IntMap   as M
 -----------------------------------------
 
 -- | Dot product of two `IntMap`s (for internal use)
-(··) :: (Num α) => IntMap α -> IntMap α -> α
-v ·· w = M.foldl' (+) 0 $ M.intersectionWith (*) v w
+(··) :: (Num α, Eq α) => IntMap α -> IntMap α -> Maybe α
+v ·· w = let x = M.foldl' (+) 0 $ M.intersectionWith (*) v w
+  in case x of
+           0 -> Nothing
+           _ -> Just x
 --    M.foldlWithKey' f 0 v
 --    where f acc 0 _ = acc
 --          f acc i x = acc + ((findWithDefault 0 i w) * x)