--- a/Math/LinearAlgebra/Sparse/IntMapUtilities.hs
+++ b/Math/LinearAlgebra/Sparse/IntMapUtilities.hs
@@ -18,6 +18,13 @@ v ·· w = case M.foldl' (+) 0 $ M.intersectionWith (*) v w of
 --    where f acc 0 _ = acc
 --          f acc i x = acc + ((findWithDefault 0 i w) * x)
 
+-- | Dot product of first `IntMap` with all `IntMap`s in second `IntMap`
+--   (for internal use)
+(···) :: (Num a, Eq a) => IntMap a -> IntMap (IntMap a) -> Maybe (IntMap a)
+r ··· m = case M.mapMaybe (r ··) m of v
+                                          | M.null v  -> Nothing
+                                          | otherwise -> Just v
+
 -- | Shifts (re-enumerates) keys of IntMap by given number
 shiftKeys :: Int -> IntMap α -> IntMap α
 shiftKeys k m = M.fromAscList [ (i+k,x) | (i,x) <- M.toAscList m ]