@laughedelic 2015-02-19
Hi @mnick! Nice idea! Thanks for the contribution :+1: 
I'll take a look at the code this weekend and merge this.
Could you also integrate your benchmarking code in the tests?