@mnick 2015-02-23
I will integrate the benchmark into the tests.

One thing I noticed since proposing the changes is that GHC doesn't automatically parallelize as well when using M.mapMaybe instead of M.map. 

I still get the better runtime performance, but this might be something to consider before merging the changes.