@aupiff 2015-06-30
Hey @mnick, what do you mean about GHC not automatically parallelizing? What are you using to achieve parallelization? Have you understood this any better?

I am testing your changes with some code I'm writing now. Will try parallel / single-threaded alternatives.